+++ src/server/raft_client.rs
@@ -188,7 +188,7 @@ impl RaftClient {
         });
 
         if counter > 0 {
-            RAFT_MESSAGE_FLUSH_COUNTER.inc_by(counter as f64);
+            RAFT_MESSAGE_FLUSH_COUNTER.inc_by(counter as i64);
         }
     }
 }
